Fix code style, unsigned RPC types and add new Memory RPC object.
 - Fix code style, to match comments from c/131311.
 - Update all unsigned types to match their primitive signed type.
 - Re-generate classes/factory and add a new Memory RPC object.

+++ android/src/com/android/tools/idea/editors/gfxtrace/controllers/ImageFetcher.java
@@ -30,13 +30,13 @@ public class ImageFetcher {
   @NotNull private Client myClient;
   private DeviceId myDeviceId;
   private CaptureId myCaptureId;
-  private Long myContextId;
+  private Integer myContextId;
 
   public ImageFetcher(@NotNull Client client) {
     myClient = client;
   }
 
-  public void prepareFetch(@NotNull DeviceId deviceId, @NotNull CaptureId captureId, @NotNull Long contextId) {
+  public void prepareFetch(@NotNull DeviceId deviceId, @NotNull CaptureId captureId, @NotNull Integer contextId) {
     ApplicationManager.getApplication().assertIsDispatchThread();
     myDeviceId = deviceId;
     myCaptureId = captureId;
