+++ android/src/com/android/tools/idea/editors/gfxtrace/rpc/Client.java
@@ -24,41 +24,22 @@ import java.util.concurrent.Future;
 
 public interface Client {
   Future<CaptureId[]> GetCaptures() throws IOException, RpcException;
-
   Future<DeviceId[]> GetDevices() throws IOException, RpcException;
-
-  Future<BinaryId> GetState(CaptureId capture, long contextId, long after) throws IOException, RpcException;
-
-  Future<HierarchyId> GetHierarchy(CaptureId capture, long contextId) throws IOException, RpcException;
-
-  Future<MemoryInfoId> GetMemoryInfo(CaptureId capture, long contextId, long after, MemoryRange rng) throws IOException, RpcException;
-
-  Future<ImageInfoId> GetFramebufferColor(DeviceId device, CaptureId capture, long contextId, long after, RenderSettings settings)
-    throws IOException, RpcException;
-
-  Future<ImageInfoId> GetFramebufferDepth(DeviceId device, CaptureId capture, long contextId, long after) throws IOException, RpcException;
-
-  Future<BinaryId> GetGlErrorCodes(DeviceId device, CaptureId capture, long contextId) throws IOException, RpcException;
-
-  Future<CaptureId> ReplaceAtom(CaptureId capture, long atomId, int atomType, Binary data) throws IOException, RpcException;
-
-  Future<TimingInfoId> GetTimingInfo(DeviceId device, CaptureId capture, long contextId, TimingMask mask) throws IOException, RpcException;
-
+  Future<BinaryId> GetState(CaptureId capture, int contextId, long after) throws IOException, RpcException;
+  Future<HierarchyId> GetHierarchy(CaptureId capture, int contextId) throws IOException, RpcException;
+  Future<MemoryInfoId> GetMemoryInfo(CaptureId capture, int contextId, long after, MemoryRange rng) throws IOException, RpcException;
+  Future<ImageInfoId> GetFramebufferColor(DeviceId device, CaptureId capture, int contextId, long after, RenderSettings settings) throws IOException, RpcException;
+  Future<ImageInfoId> GetFramebufferDepth(DeviceId device, CaptureId capture, int contextId, long after) throws IOException, RpcException;
+  Future<BinaryId> GetGlErrorCodes(DeviceId device, CaptureId capture, int contextId) throws IOException, RpcException;
+  Future<CaptureId> ReplaceAtom(CaptureId capture, long atomId, short atomType, Binary data) throws IOException, RpcException;
+  Future<TimingInfoId> GetTimingInfo(DeviceId device, CaptureId capture, int contextId, TimingMask mask) throws IOException, RpcException;
   Future<AtomStream> ResolveAtomStream(AtomStreamId id) throws IOException, RpcException;
-
   Future<Binary> ResolveBinary(BinaryId id) throws IOException, RpcException;
-
   Future<Capture> ResolveCapture(CaptureId id) throws IOException, RpcException;
-
   Future<Device> ResolveDevice(DeviceId id) throws IOException, RpcException;
-
   Future<Hierarchy> ResolveHierarchy(HierarchyId id) throws IOException, RpcException;
-
   Future<ImageInfo> ResolveImageInfo(ImageInfoId id) throws IOException, RpcException;
-
   Future<MemoryInfo> ResolveMemoryInfo(MemoryInfoId id) throws IOException, RpcException;
-
   Future<Schema> ResolveSchema(SchemaId id) throws IOException, RpcException;
-
   Future<TimingInfo> ResolveTimingInfo(TimingInfoId id) throws IOException, RpcException;
 }
